+++ lib/tiny_tds/result.rb
@@ -1,5 +1,11 @@
 module TinyTds
   class Result
+    attr_reader :affected_rows, :fields, :rows, :return_code
+
     include Enumerable
+
+    def each(&bk)
+      rows.each(&bk)
+    end
   end
 end